refactor(cli): dissolve pkg/cliflag, move Enum to cmd and Output to pkg/generator
Move the generic Enum[T] pflag.Value and its flag infrastructure from
pkg/cliflag into the cmd package, where it is used. Introduce
pkg/generator as the home for output-rendering types, starting with the
Output enum and its constants.

- pkg/cliflag deleted
- cli/cmd/enum.go: Enum[T], NewEnum, ErrInvalidValue, NewOutputFlag()
- cli/pkg/generator/output.go: Output type and constants (no constructor)
- cli/cmd/flags.go: call NewOutputFlag() directly
- cli/cmd/enum_test.go: adapted from the former cliflag/output_test.go
