+++ cli/cmd/enum_test.go
@@ -1,15 +1,17 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2026 NVIDIA Corporation
 
-package cliflag
+package cmd
 
 import (
 	"errors"
 	"testing"
+
+	"github.com/run-ai/karta/cli/pkg/generator"
 )
 
 func TestOutputSetValid(t *testing.T) {
-	o := NewOutput()
+	o := NewOutputFlag()
 	for _, v := range o.Allowed() {
 		if err := o.Set(v); err != nil {
 			t.Errorf("Set(%q) returned error: %v", v, err)
@@ -24,7 +26,7 @@ func TestOutputSetValid(t *testing.T) {
 }
 
 func TestOutputSetInvalid(t *testing.T) {
-	o := NewOutput()
+	o := NewOutputFlag()
 	err := o.Set("bogus")
 	if err == nil {
 		t.Fatal("expected error for invalid output value, got nil")
@@ -33,13 +35,13 @@ func TestOutputSetInvalid(t *testing.T) {
 		t.Errorf("error is not ErrInvalidValue: %v", err)
 	}
 	// An invalid Set must not mutate the value away from its default.
-	if o.Get() != OutputTable {
+	if o.Get() != generator.OutputTable {
 		t.Errorf("value changed after failed Set: %q", o.String())
 	}
 }
 
 func TestOutputType(t *testing.T) {
-	if got := NewOutput().Type(); got != "output" {
+	if got := NewOutputFlag().Type(); got != "output" {
 		t.Errorf("Type() = %q, want %q", got, "output")
 	}
 }
